requirements.txt
aiogram==2.25.1
python-dotenv==0.21.1
g4f==0.1.3.5